changed tests
