+++ cypress/e2e/settings/data-import/delete-an-existing-job-profile.cy.js
@@ -1,27 +1,72 @@
+import { FOLIO_RECORD_TYPE } from '../../../support/constants';
+import { Permissions } from '../../../support/dictionary';
+import NewActionProfile from '../../../support/fragments/data_import/action_profiles/newActionProfile';
 import JobProfileView from '../../../support/fragments/data_import/job_profiles/jobProfileView';
 import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
 import NewJobProfile from '../../../support/fragments/data_import/job_profiles/newJobProfile';
+import {
+  ActionProfiles as SettingsActionProfiles,
+  FieldMappingProfiles as SettingsFieldMappingProfiles,
+} from '../../../support/fragments/settings/dataImport';
+import NewFieldMappingProfile from '../../../support/fragments/settings/dataImport/fieldMappingProfile/newFieldMappingProfile';
 import SettingsMenu from '../../../support/fragments/settingsMenu';
+import Users from '../../../support/fragments/users/users';
 import getRandomPostfix from '../../../support/utils/stringTools';
 
 describe('Data Import', () => {
   describe('Settings', () => {
-    const jobProfileName = `C2334 autotest job profile ${getRandomPostfix()}`;
-    const calloutMessage = `The job profile "${jobProfileName}" was successfully deleted`;
+    let user;
+    const mappingProfile = {
+      name: `C2334 autotest mapping profile_${getRandomPostfix()}`,
+      typeValue: FOLIO_RECORD_TYPE.INSTANCE,
+    };
+    const actionProfile = {
+      name: `C2334 autotest action profile_${getRandomPostfix()}`,
+      action: 'CREATE',
+      folioRecordType: 'INSTANCE',
+    };
+    const jobProfile = {
+      profileName: `C2334 autotest job profile ${getRandomPostfix()}`,
+    };
+    const calloutMessage = `The job profile "${jobProfile.profileName}" was successfully deleted`;
 
     before('Create test data and login', () => {
       cy.getAdminToken();
-      NewJobProfile.createJobProfileWithoutLinkedProfilesViaApi(jobProfileName);
+      NewFieldMappingProfile.createInstanceMappingProfileViaApi(mappingProfile).then(
+        (mappingProfileResponse) => {
+          NewActionProfile.createActionProfileViaApi(
+            actionProfile,
+            mappingProfileResponse.body.id,
+          ).then((actionProfileResponse) => {
+            NewJobProfile.createJobProfileWithLinkedActionProfileViaApi(
+              jobProfile.profileName,
+              actionProfileResponse.body.id,
+            );
+          });
+        },
+      );
 
-      cy.loginAsAdmin();
-      cy.visit(SettingsMenu.jobProfilePath);
+      cy.createTempUser([Permissions.settingsDataImportEnabled.gui]).then((userProperties) => {
+        user = userProperties;
+
+        cy.login(user.username, user.password);
+        cy.visit(SettingsMenu.jobProfilePath);
+      });
+    });
+
+    after('Delete test data', () => {
+      cy.getAdminToken().then(() => {
+        SettingsActionProfiles.deleteActionProfileByNameViaApi(actionProfile);
+        SettingsFieldMappingProfiles.deleteMappingProfileByNameViaApi(mappingProfile);
+        Users.deleteViaApi(user.userId);
+      });
     });
 
     it(
       'C2334 Delete an existing job profile (folijet)',
       { tags: ['extendedPath', 'folijet'] },
       () => {
-        JobProfiles.search(jobProfileName);
+        JobProfiles.search(jobProfile.profileName);
         JobProfileView.delete();
         NewJobProfile.checkCalloutMessage(calloutMessage);
         JobProfiles.verifyJobProfileAbsent();
